Created De-jamming Mechanism
I created a De-jamming mechanism that should run both intake and outtake the opposite way they normally move, this should get a stuck or broken note out of the robot so we can continue to pick up and shoot normal notes. There is one error that I know of, the Autonomous Command has been removed from Robot container and that has led to an error in the Robot file.